+++ lib/integral/middleware/alias_router.rb
@@ -1,16 +1,17 @@
 module Integral
   # Integral Middleware
   module Middleware
-    # Handles dynamic page routing.
-    # Checks all GET requests to see if the PATH matches an Integral::Page path
-    # If a match is found the PATH is rewritten from human readable into something Rails
-    # router understands. i.e. '/company/who-we-are' -> /pages/12
-    class PageRouter
+    # Handles dynamic page paths and category slug routing.
+    #
+    # Checks all GET requests to see if the PATH matches an Integral::Page path or category slug
+    # If a match is found the PATH is rewritten from human readable into a Rails route i.e.
+    # '/company/who-we-are' -> /pages/12
+    class AliasRouter
       def initialize(app)
         @app = app
       end
 
-      # Handles dynamic Integral::Page routing.
+      # Handles rewriting the path to something the Rails router will understand
       def call(env)
         request = Rack::Request.new(env)
 
@@ -22,7 +23,7 @@ module Integral
         return @app.call(env) if request.path_info.starts_with?(backend_path)
 
         # Rewrites path if the request linked to an Integral::Page or Integral::Category
-        rewrite_path(env, request.path_info)
+        process_path(env, request)
 
         @app.call(env)
       end
@@ -30,7 +31,6 @@ module Integral
       private
 
       # Checks to see if path is linked to a page
-      # TODO: Rather than hitting the DB on each request here a Redis solution could be implemented
       #
       # @param path [String] Path the request is linked to
       #
@@ -58,22 +58,28 @@ module Integral
       # i.e. '/company/who-we-are' -> /pages/12
       #
       # @param env [Hash] Environment of the request
-      # @param path [String] Path the request is linked to
+      # @param request [Hash] the request
       #
       # @return [String] Path Rails needs to link the request to the correct Integral::Page record
-      def rewrite_path(env, path)
+      def process_path(env, request)
+        path = request.path_info
         locale = Integral.frontend_locales.find { |locale| path.starts_with?("/#{locale}/") || path == "/#{locale}" } || Integral.frontend_locales.first if Integral.multilingual_frontend?
 
         page_id = page_identifier(path)
         if page_id
-          env['PATH_INFO'] = localized_path("/pages/#{page_id}", locale)
+          rewrite_path(env, request, localized_path("/pages/#{page_id}", locale))
         else
-          category_id = category_identifier(path, locale)
-
-          env['PATH_INFO'] = localized_path("/#{Integral.blog_namespace}/categories/#{category_id}", locale) if category_id
+          if category_id = category_identifier(path, locale)
+            rewrite_path(env, request, localized_path("/#{Integral.blog_namespace}/categories/#{category_id}", locale))
+          end
         end
       rescue StandardError => e
         handle_rewrite_error(e)
+      end
+
+      def rewrite_path(env, request, new_path)
+        request.update_param(:integral_original_path, env['PATH_INFO'])
+        env['PATH_INFO'] = new_path
       end
 
       def localized_path(path, locale)